packages/rag/requirements.txt
annotated-types==0.7.0
anyio==4.12.1
certifi==2026.2.25
charset-normalizer==3.4.4
click==8.3.1
docker==7.1.0
fastapi==0.115.12
h11==0.16.0
httptools==0.7.1
idna==3.11
numpy==2.2.4
psycopg==3.2.6
psycopg-binary==3.2.6
pydantic==2.12.5
pydantic_core==2.41.5
python-dotenv==1.0.1
PyYAML==6.0.3
requests==2.32.3
starlette==0.46.2
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.34.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==16.0
